requirements.txt
click==8.1.3
email_validator==1.3.1
Flask==2.2.3
Flask_Login==0.6.2
Flask_Migrate==4.0.4
flask_sqlalchemy==3.0.3
Flask_WTF==1.1.1
pytest==7.2.2
python-dotenv==1.0.0
SQLAlchemy==2.0.7
stripe==5.3.0
Werkzeug==2.2.3
WTForms==3.0.1
